fix(menu.py): Display apps without metadata.json
If an app directory is missing metadata.json or if metadata.json is
invalid/cannot be parsed correctly, don't just print an error, also add
a menu entry based on what we'd hope to be correct information.  This is
obviously best effort only and there is no guarantee that it will work.

Signed-off-by: Rahix <rahix@rahix.de>