fix(ws2812): Fix first write making the first pixel green
The first epic_ws2812_write() call will set the first pixel to 0x008000
(bright green).  This is caused by the GPIO line being pulled down on
mode-setting (epic_gpio_set_pin_mode).  Wait before writing the values
to reset the bus and thus properly set the pixels to the correct colors
on first write.

The first LED will still flash green for a very short time.  I don't
know if there is a workaround for this ...

Signed-off-by: Rahix <rahix@rahix.de>